FIX Cleanup App.tsx from previous merge :
Description:
	The previous merge was having only one route,
	It was implemented like this intentionnaly for helping to have a easy theme test (light and dark mode)
	now this page will be available on "/first_step_theme" route except registration step
