+++ front/app/src/Views/BaseViewFromVitejs/BaseViewFromViteJs.tsx
@@ -3,6 +3,7 @@ import LogoBase from "./components/LogoBase";
 import TitleBase from "./components/TitleBase";
 import CardBase from "./components/CardBase";
 import DocInfoBase from "./components/DocInfoBase";
+import TestComponent from "./components/TestComponent";
 
 import "./styles/Base.css";
 
@@ -14,6 +15,7 @@ const	BaseViewFromViteJs = () =>
 			<TitleBase />
 			<CardBase />
 			<DocInfoBase />
+			<TestComponent />
 		</>
 	);
 };
